Create YHWHcoin.sol
YHVHcoin is a sovereign Cryptoactive. Issued by the Community Economy Bank of the LA PAZ Socialist Community through the TRON System Block Chain.

The YHWH Cryptoactive Tokens issued to achieve the decentralization of the Community Economic system of the Bank's financial system Centrals Opening the execution and development of our platform for Crypto Assets as a Community Economic Exchange System in function of the insurgency of the Popular Power.

Reservation of denominations or names of the different forms of organizations of the Popular Power of the Organized Community of the BARINAS CAPITAL COMMUNAL DISTRICT, Constituent of Agrarian and Socialist Communes is described below [square brackets with image of Claw], established in RESOLUTION ES-010 OF THE SOCIALIST AXIS, Duly Protocolized to the Public Registry Office with Notarial functions under No. 03, Folios 18 to 30, First (1st) Protocol, Second (2nd) main Volume, Fourth quarter of the year two thousand sixteen. "The gerarchic symbol between points and square brackets". synthesize the titling chain of more than four thousand (4,000) Base Council of the Popular Power. Establishing itself through this title chain of Blocks Articulation of the community in general with the INTI-EJE Presidential Commission SOCIALIST THE NEW GEOMETRY OF POWER. Articulation of Fact of Law and Justice which prescribes the social and geographical scope as a system of communal aggregation.

it is foreshadowed that the forms of construction of supremacy will only be achieved accelerating the change of the ECONOMIC SYSTEM, transcending the Petroleum Rentista Model, to the Productive Economic Model, giving way to a more egalitarian and just society, sustaining in the role of the Social and Democratic State, of law and Justice, in order to continue advancing in full satisfaction of the basic needs for the life of our people: Food, Water, Electricity, Housing and housing, public transport, health, education, public safety, access to culture, communication, free, science and technology, sports, healthy recreation and decent work liberated and liberating, turning VENEZUELA into a Social and Political Power Country within the nascent power of Latin America and the Caribbean, and guarantees the execution of peace zones in our America, contribute to the development of a New International GeoEconomy, in which it takes body a Multicentric and Pluripolar World, which allows to achieve the balance of the Universe and We need Planetary Peace. Preserve life on the planet and preserve the Human Species, translating in the need for the Execution of an ECOSOCIALIST Productive Economic Model, Based on a harmonic relationship between man and nature, and the rational and optimal use and exploitation of Natural resources, representing Natural processes and cycles.

The advancement of new technologies allows for the implementation of reliable information systems, that will allow the transition process of our society and that of bureaucratic democratism be much more flexible, reduce public expenses and achieve the ultimate irruption of the state Bureaucratist, our platform aims to integrate the different systems that it currently manages the public administration in Venezuela as a pilot plan, for more reliable systems in the use of Resources Economical and its equitable and equal distribution of it, our platform will positively impact the environment Environment, significantly reduces the ECOCIDIO, reduces the operating time and connectivity of the Conventional banking platforms, reduce energy consumption, obtain real-time data, it allows to achieve a better management control over the use of collective public resources eradicating the correction, the reduction of budget in the electoral processes, due to the reduction of consumption of natural materials; as paper, plastics among others at 0%, with the implementation of digital contracts adapted to give greater guarantee to the consultative processes. the elimination of contaminating stationery that generates lush sums of public expenses in collection, recycling, transportation, the eradication of processes obsolete and a step towards raising awareness of social systems.